src/requirements.txt
streamlit
google-genai
pinecone
pymupdf
python-dotenv
Pillow
openai
